# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Open-UAV-Telemetry-Bridge API
  description: |
    REST API for the Open-UAV-Telemetry-Bridge (OUTB) gateway.
    Provides access to drone telemetry data, system status, configuration management,
    alerts, geofences, and real-time WebSocket streaming.
  version: 0.4.0
  contact:
    name: OUTB Project
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://your-server.example.com
    description: Production server (HTTPS)
...
paths:
...
  /api/v1/status:
    get:
      tags:
        - Status
      summary: Get gateway status
      description: Returns gateway version, uptime, and statistics
      security:
        - bearerAuth: []
        - {}
      responses:
        '200':
          description: Gateway status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/drones:
    get:
      tags:
        - Drones
      summary: List all drones
      description: Returns all currently known drone states
      security:
        - bearerAuth: []
        - {}
      responses:
        '200':
          description: List of drone states
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DronesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/drones/{deviceID}:
    get:
      tags:
        - Drones
      summary: Get drone by ID
      description: Returns the state of a specific drone
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: deviceID
          in: path
          required: true
          schema:
            type: string
          description: Drone device ID
          example: drone-001
      responses:
        '200':
          description: Drone state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DroneState'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/drones/{deviceID}/track:
    get:
      tags:
        - Tracks
      summary: Get drone track
      description: Returns historical trajectory points for a drone
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: deviceID
          in: path
          required: true
          schema:
            type: string
          description: Drone device ID
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
          description: Maximum number of points to return
        - name: since
          in: query
          schema:
            type: integer
            format: int64
          description: Unix timestamp (ms) - only return points after this time
      responses:
        '200':
          description: Track points
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          description: Track storage is disabled
    delete:
      tags:
        - Tracks
      summary: Clear drone track
      description: Deletes all historical trajectory points for a drone
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: deviceID
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Track cleared successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          description: Track storage is disabled
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: Login
      description: Authenticate and receive a JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
...
  /api/v1/auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Returns information about the authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/config:
    get:
      tags:
        - Configuration
      summary: Get configuration
      description: Returns the current gateway configuration (sensitive fields redacted)
      security:
        - bearerAuth: []
        - {}
      responses:
        '200':
          description: Current configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppConfig'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/logs:
    get:
      tags:
        - Logs
      summary: Get logs
      description: Returns recent log entries from the buffer
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: level
          in: query
          schema:
            type: string
            enum: [debug, info, warn, error]
          description: Minimum log level to return
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
          description: Maximum number of entries to return
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Number of entries to skip
      responses:
        '200':
          description: Log entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags:
        - Logs
      summary: Clear logs
      description: Clears all log entries from the buffer
      security:
        - bearerAuth: []
        - {}
      responses:
        '204':
          description: Logs cleared
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/logs/stream:
    get:
      tags:
        - Logs
      summary: Stream logs (SSE)
      description: Server-Sent Events stream of real-time log entries
      security:
        - bearerAuth: []
        - {}
      responses:
        '200':
          description: SSE stream of log entries
          content:
            text/event-stream:
              schema:
                type: string
  /api/v1/alerts:
    get:
      tags:
        - Alerts
      summary: List alerts
      description: Returns all alerts with optional filtering
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: severity
          in: query
          schema:
            type: string
            enum: [info, warning, critical]
        - name: acknowledged
          in: query
          schema:
            type: boolean
        - name: device_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Alert list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags:
        - Alerts
      summary: Clear alerts
      description: Clears all alerts
      security:
        - bearerAuth: []
        - {}
      responses:
        '204':
          description: Alerts cleared
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/alerts/{id}:
    get:
      tags:
        - Alerts
      summary: Get alert by ID
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Alert details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/alerts/{id}/ack:
    post:
      tags:
        - Alerts
      summary: Acknowledge alert
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Alert acknowledged
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/alerts/rules:
    get:
      tags:
        - Alerts
      summary: List alert rules
      security:
        - bearerAuth: []
        - {}
      responses:
        '200':
          description: Alert rules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRulesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - Alerts
      summary: Create alert rule
      security:
        - bearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertRule'
      responses:
        '201':
          description: Rule created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRule'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/geofences:
    get:
      tags:
        - Geofences
      summary: List geofences
      security:
        - bearerAuth: []
        - {}
      responses:
        '200':
          description: Geofence list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeofencesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - Geofences
      summary: Create geofence
      security:
        - bearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Geofence'
      responses:
        '201':
          description: Geofence created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Geofence'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/geofences/{id}:
    get:
      tags:
        - Geofences
      summary: Get geofence by ID
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Geofence details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Geofence'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags:
        - Geofences
      summary: Update geofence
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Geofence'
      responses:
        '200':
          description: Geofence updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Geofence'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags:
        - Geofences
      summary: Delete geofence
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Geofence deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/geofences/breaches:
    get:
      tags:
        - Geofences
      summary: List geofence breaches
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: geofence_id
          in: query
          schema:
            type: string
        - name: device_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Breach list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BreachesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags:
        - Geofences
      summary: Clear breaches
      security:
        - bearerAuth: []
        - {}
      responses:
        '204':
          description: Breaches cleared
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/ws:
    get:
      tags:
        - WebSocket
      summary: WebSocket connection
      description: |
        Establishes a WebSocket connection for real-time drone state updates.

        Message types received:
        - `state_update`: Drone state changed
        - `drone_online`: New drone connected
        - `drone_offline`: Drone disconnected
      security:
        - bearerAuth: []
        - {}
      responses:
        '101':
          description: Switching to WebSocket protocol
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/auth/login
  responses:
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    StatusResponse:
      type: object
      properties:
        version:
          type: string
          example: 0.4.0-dev
        uptime_seconds:
          type: integer
          format: int64
          example: 3600
        adapters:
          type: array
          items:
            $ref: '#/components/schemas/AdapterStatus'
        publishers:
          type: array
          items:
            type: string
          example: [mqtt, gb28181]
        stats:
          $ref: '#/components/schemas/Stats'
...
    DroneState:
      type: object
      required:
        - device_id
        - timestamp
        - protocol_source
        - location
        - attitude
        - velocity
        - status
      properties:
        device_id:
          type: string
          example: drone-001
        timestamp:
          type: integer
          format: int64
          description: Unix timestamp in milliseconds
          example: 1705833600000
        protocol_source:
          type: string
          enum: [mavlink, dji_mobile_sdk]
          example: mavlink
        location:
          $ref: '#/components/schemas/Location'
        attitude:
          $ref: '#/components/schemas/Attitude'
        velocity:
          $ref: '#/components/schemas/Velocity'
        status:
          $ref: '#/components/schemas/Status'
...
    DronesResponse:
      type: object
      properties:
        count:
          type: integer
          example: 2
        drones:
          type: array
          items:
            $ref: '#/components/schemas/DroneState'
...
    TrackResponse:
      type: object
      properties:
        device_id:
          type: string
        count:
          type: integer
        points:
          type: array
          items:
            $ref: '#/components/schemas/TrackPoint'
        total_size:
          type: integer
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          format: password
          example: admin123
...